8_rl_cookbook/sama_rl/requirements.txt
torch>=2.1.0
transformers>=4.30.0
datasets>=2.0.0
trl>=0.18.0
wandb>=0.15.0
accelerate>=0.20.0
peft>=0.4.0
typing_extensions>=4.8.0
PyYAML>=6.0
sagemaker>=2.100.0
boto3>=1.20.0
